move deleteAccount() to AccountManager
