+++ ts/components/session/LeftPaneSettingSection.tsx
@@ -14,6 +14,7 @@ import { SessionSearchInput } from './SessionSearchInput';
 import { SessionSettingCategory } from './settings/SessionSettings';
 import { DefaultTheme } from 'styled-components';
 import { LeftPaneSectionHeader } from './LeftPaneSectionHeader';
+import { AccountManager } from '../../util';
 
 interface Props {
   settingsCategory: SessionSettingCategory;
@@ -179,7 +180,7 @@ export class LeftPaneSettingSection extends React.Component<Props, State> {
       title,
       message,
       messageSub,
-      resolve: window.deleteAccount,
+      resolve: AccountManager.deleteAccount,
       okTheme: 'danger',
     });
   }
